* master:
  Changes to Gaussian class and handler.
  Fix m2r2 issue. See https://github.com/sphinx-doc/sphinx/issues/8395
  Added version requirement of msgpack>=0.6.0 in requirement.txt
  m2r fix.
  neuralnetwork module compatible with new pytorch versions.

+++ ml4chem/atomistic/models/neuralnetwork.py
@@ -787,7 +787,7 @@ class train(DeepLearningTrainer):
         grads = sum(grads)
 
         for index, param in enumerate(model.parameters()):
-            param.grad = torch.tensor(grads[index])
+            param.grad = torch.tensor(grads[index], dtype=torch.float)
 
         del accumulation
         del grads
